pyproject.toml
[project]
name = "video2srt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.0",
    "moviepy>=2.2.1",
    "pyyaml>=6.0.3",
    "yt-dlp>=2026.3.17",
    "google-genai>=1.73.1",
]
